requirements.txt
streamlit
pandas
python-dotenv
openai  # For Together API access
plotly  # For interactive visualizations
